libtest_utils: Moved to own directories
