+++ include/test_utils/context-util.h
@@ -2,7 +2,7 @@
 #define CONTEXT_UTIL_H
 
 #include "sapi/tpm20.h"
-#include "test-options.h"
+#include "test_utils/test-options.h"
 
 /**
  * functions to setup TCTIs and SAPI contexts  using data from the common
